refact : separate renderVnode related functions in client
